requirements.txt
numpy==1.23.5
opencv-python
pandas
Flask
Flask-SocketIO==5.3.6
python-engineio==4.9.0
python-socketio==5.11.1
keyboard
faiss-cpu
pyserial
deepface
tensorflow==2.10.1
keras==2.10.0
setuptools==65.5.0